+++ src/SecurityConsultantCore/Domain/Basic/XYZLocation.cs
@@ -10,8 +10,8 @@
 
         public XYZ Location { get; }
         public T Obj { get; }
-        public int X => Location.X;
-        public int Y => Location.Y;
+        public double X => Location.X;
+        public double Y => Location.Y;
         public int Z => Location.Z;
     }
 }
